fixed invalid json on flag delete

+++ frontend_rework/src/lib/flags.ts
@@ -1,15 +1,16 @@
 import { api } from '$lib/api'
 
 export async function deleteFlags(flags: any[], chall_id:any){
-  const list = Array.from(flags ?? [])
-    .map((flag) =>
+  const requests = flags.map(f =>{
+    if (f.flag !== "")
       api<any>('/flags', {
         method: 'DELETE',
         headers: { "content-type": "application/json" },
-        body: JSON.stringify({ flag, chall_id }), // let `api` set JSON headers & stringify
+        body: JSON.stringify({ 'flag': f.flag, chall_id }),
       })
-    );
-  return list;
+  }
+  );
+  return requests;
 }
 
 export async function createFlags(flags: any[], chall_id:any){
